Refactor instance json into a separate service
